requirements.txt
google-genai==1.10.0
google-generativeai==0.8.4